refactor: complete restructuring of sshnpd.dart
